test_offscreencanvas_basic_webgl.html

<!DOCTYPE HTML>
<html>
<head>
<title>WebGL in OffscreenCanvas</title>
<meta name="viewport" content="width=device-width, user-scalable=no, initial-scale=1">
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
</head>
<body>
<canvas id="c" width="64" height="64"></canvas>
<a href="index.html">Index</a>
<div id="info"></div>
<script>

SimpleTest.waitForExplicitFinish();

function runTest() {

  var htmlCanvas = document.getElementById("c");
  var worker = new Worker("js/offscreencanvas.js");

  ok(htmlCanvas, "Should have HTML canvas element");
  ok(worker, "Web worker successfully created");

  worker.onmessage = function(evt) {
    var msg = evt.data || {};
    if (msg.type == "test") {
      ok(msg.result, msg.name);
    }
    if (msg.type == "finish") {
      worker.terminate();
      SimpleTest.finish();
    }
  }

  ok(htmlCanvas.transferControlToOffscreen, "HTMLCanvasElement has transferControlToOffscreen function");

  var offscreenCanvas = htmlCanvas.transferControlToOffscreen();
  ok(offscreenCanvas, "Expected transferControlToOffscreen to succeed");

  worker.postMessage({test: 'webgl', canvas: offscreenCanvas}, [offscreenCanvas]);
}


SpecialPowers.pushPrefEnv({'set': [
  ['gfx.offscreencanvas.enabled', true],
  ['webgl.force-enabled', true],
]}, runTest);
</script>
</body>
</html>
